guardado 30-12-25

+++ app/pages/clientes/actions/delete-client-action.ts
@@ -5,9 +5,15 @@ import { createClient } from "@/utils/supabase/server";
 export const deleteClientAction = async (clientId: string) => {
     try {
         const supabase = await createClient();
+        const { error: deletePaymentsError } = await supabase.from('payments').delete().eq('client_id', clientId);
+        if (deletePaymentsError) {
+            console.error("Error al eliminar los pagos del cliente:", deletePaymentsError);
+            throw new Error (deletePaymentsError.message);
+        }
         const { data, error } = await supabase.from('clients').delete().eq('id_client', clientId);
         if (error) {
-            throw error;
+            console.error("Error al eliminar el cliente:", error);
+            throw new Error (error.message);
         }
         return data;
     } catch (error) {
